requirements.txt
django==4.1.2
djangorestframework==3.14.0
django-allauth==0.52.0
django-extensions==3.2.1
django-filter==22.1
dj-rest-auth==2.2.5
factory-boy==3.2.1
psycopg2-binary==2.9.5